[clients_gui] better fix for warning in LocationEditorToolBar
For some reason, Qt does not want my size_t in its slot, so be it.

+++ src/libraries/clients_gui/LocationEditorToolbar.h
@@ -54,7 +54,7 @@ private slots:
     void Goto();
     void AddParamPoint();
     void CreateBookmark();
-    void GotoBookmark( size_t index );
+    void GotoBookmark( int index );
     void ClearBookmarks();
     //@}
